pool-d: decomp 13 more resource utility functions — arithmetic, bitfield, no-ops
fun_hard_d_034.cpp additions:
- 3 arithmetic operators (add, subtract, divide-by-200)
- 5 bitfield extractors/masks (3x 4-bit extract, 2x clear-lower-32)
- 3 empty no-op functions (optnone stubs)
- 2 conditional resource init functions (init guard + handle allocator)

fun_hard_d_031.cpp: fix FUN_71039521d0 extern return type (u32 -> u64)

+++ src/app/fun_hard_d_031.cpp
@@ -23,7 +23,7 @@ extern "C" void  FUN_7103951cd0(u64);
 extern "C" void  FUN_7103951d20(u64);
 extern "C" void  FUN_7103951de0(u64);
 extern "C" void  FUN_7103953390(u64);
-extern "C" u32   FUN_71039521d0(u32);
+extern "C" u64   FUN_71039521d0(u32);
 extern "C" u32   FUN_7103953450(u64);
 extern "C" void  FUN_71037c90e0(u64, u64);
 extern "C" void  FUN_7103958740(u64);
